untagging people from posts; updating 'transactional' methods; changing access modifiers in services; ensuring validation of post's owner consistency through all services; other refactoring and renaming

+++ src/main/java/com/example/travelleronline/posts/PostService.java
@@ -7,7 +7,7 @@ import com.example.travelleronline.general.exceptions.UnauthorizedException;
 import com.example.travelleronline.posts.dtos.PostEditDTO;
 import com.example.travelleronline.posts.dtos.PostFilterDTO;
 import com.example.travelleronline.posts.dtos.PostDTO;
-import com.example.travelleronline.reactions.LikesDislikesDTO;
+import com.example.travelleronline.reactions.dto.LikesDislikesDTO;
 import com.example.travelleronline.reactions.toPost.PostReaction;
 import com.example.travelleronline.posts.dtos.PostCreationDTO;
 import com.example.travelleronline.users.User;
@@ -18,7 +18,6 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -27,18 +26,18 @@ import java.util.stream.Collectors;
 @Service
 public class PostService extends MasterService {
 
-    public static final int LONGITUDE_MIN_VALUE = -180;
-    public static final int LONGITUDE_MAX_VALUE = 180;
-    public static final int LATITUDE_MIN_VALUE = -90;
-    public static final int LATITUDE_MAX_VALUE = 90;
-    public static final int POST_DESCRIPTION_LENGTH_MAX = 500;
-    public static final int POST_TITLE_LENGTH_MIN = 3;
-    public static final int POST_TITLE_LENGTH_MAX = 100;
+    private static final int LONGITUDE_MIN_VALUE = -180;
+    private static final int LONGITUDE_MAX_VALUE = 180;
+    private static final int LATITUDE_MIN_VALUE = -90;
+    private static final int LATITUDE_MAX_VALUE = 90;
+    private static final int POST_DESCRIPTION_LENGTH_MAX = 500;
+    private static final int POST_TITLE_LENGTH_MIN = 3;
+    private static final int POST_TITLE_LENGTH_MAX = 100;
 
     @Autowired
     private PostDAO dao;
 
-    public PostCreationDTO createPost(PostCreationDTO dto, int uid) {
+    PostCreationDTO createPost(PostCreationDTO dto, int uid) {
         Category c = validatePost(dto);
         User u = getVerifiedUserById(uid);
         Post p = modelMapper.map(dto,Post.class);
@@ -51,23 +50,16 @@ public class PostService extends MasterService {
         return dto;
     }
 
-    public PostDTO getPost(int pid) {
+    PostDTO getPost(int pid) {
         return modelMapper.map(getPostById(pid), PostDTO.class);
     }
 
-    public void deletePostById(int pid, int uid) {
-        validateDeletionOfPost(pid, uid);
+    void deletePostById(int pid, int uid) {
+        validatePostOwner(pid, uid);
         postRepository.deleteById(pid);
     }
 
-    private void validateDeletionOfPost(int pid, int uid) {
-        Post post = getPostById(pid);
-        if (post.getOwner().getUserId() != uid) {
-            throw new UnauthorizedException("You must be the post owner to delete this post.");
-        }
-    }
-
-    public void editPost(int pid, PostEditDTO dto, int uid) {
+    void editPost(int pid, PostEditDTO dto, int uid) {
         Post existingPost = validatePostOwner(pid, uid);
         Category c = validatePost(dto);
         existingPost.setCategory(c);
@@ -92,8 +84,10 @@ public class PostService extends MasterService {
     private void validateLocation(double locationLatitude, double locationLongitude) {
         if (locationLongitude < LONGITUDE_MIN_VALUE || locationLongitude > LONGITUDE_MAX_VALUE ||
             locationLatitude < LATITUDE_MIN_VALUE || locationLatitude > LATITUDE_MAX_VALUE) {
-            throw new BadRequestException("Location latitude must be a number between " + LATITUDE_MIN_VALUE + " and " + LATITUDE_MAX_VALUE +
-                    " and the longitude between " + LONGITUDE_MIN_VALUE + " and " + LONGITUDE_MAX_VALUE + ".");
+            throw new BadRequestException("Location latitude must be a number between " +
+                    LATITUDE_MIN_VALUE + " and " + LATITUDE_MAX_VALUE +
+                    " and the longitude between " + LONGITUDE_MIN_VALUE +
+                    " and " + LONGITUDE_MAX_VALUE + ".");
         }
     }
 
@@ -107,50 +101,61 @@ public class PostService extends MasterService {
 
     private void validateDescription(String desc) {
         if (desc.length() > POST_DESCRIPTION_LENGTH_MAX) {
-            throw new BadRequestException("Description must not be over " + POST_DESCRIPTION_LENGTH_MAX + " letters.");
+            throw new BadRequestException("Description must not be over " +
+                    POST_DESCRIPTION_LENGTH_MAX + " letters.");
         }
     }
 
     private void validateTitle(String title) {
-        if (title == null || title.equals("null")) {
-            throw new BadRequestException("Title can not be null.");
+        if (title == null || title.isBlank()) {
+            throw new BadRequestException("Title can not be blank.");
         }
-        if (title.length() < POST_TITLE_LENGTH_MIN || title.isBlank() || title.length() > POST_TITLE_LENGTH_MAX) {
-            throw new BadRequestException("Title must be between " + POST_TITLE_LENGTH_MIN + " and " + POST_TITLE_LENGTH_MAX + " letters");
+        if (title.length() < POST_TITLE_LENGTH_MIN || title.length() > POST_TITLE_LENGTH_MAX) {
+            throw new BadRequestException("Title must be between " + POST_TITLE_LENGTH_MIN +
+                    " and " + POST_TITLE_LENGTH_MAX + " letters.");
         }
     }
 
-    @Transactional
-    public void tagUserToPost(int pid, int uid, int sessionUserId) {
+    // untags after second visit
+    void tagUserToPost(int pid, int uid, int sessionUserId) {
         Post p = validatePostOwner(pid,sessionUserId);
         User u = getVerifiedUserById(uid);
-        p.getTaggedUsers().add(u);
-        u.getTaggedInPosts().add(p);
-        postRepository.save(p);
+        if (u.getTaggedInPosts().contains(p)) {
+            u.getTaggedInPosts().remove(p);
+        }
+        else {
+            u.getTaggedInPosts().add(p);
+        }
         userRepository.save(u);
     }
 
-    public List<PostFilterDTO> filterPosts(String searchBy, String value, String orderBy,
+    List<PostFilterDTO> filterPosts(String searchBy, String searchVaule, String orderBy,
                                            int pageNumber, int rowsNumber) {
-        return dao.filterPosts(searchBy, value, orderBy, pageNumber, rowsNumber);
+        if (searchBy.equals("title")) {
+            StringBuilder builder = new StringBuilder();
+            searchVaule = searchVaule.replace("_", " ");
+            searchVaule = builder.append("%").append(searchVaule).append("%").toString();
+        }
+        return dao.filterPosts(searchBy, searchVaule, orderBy, pageNumber, rowsNumber);
     }
 
-    public List<PostFilterDTO> getPostsByCategory(String category, int pageNumber, int rowsNumber) {
+    List<PostFilterDTO> getPostsByCategory(String category, int pageNumber, int rowsNumber) {
         Category c = validateCategory(category);
         Pageable page = PageRequest.of(pageNumber, rowsNumber, Sort.by("dateOfUpload"));
         List<Post> posts = postRepository.findAllByCategory(c, page);
         return posts.stream().map(p -> modelMapper.map(p,PostFilterDTO.class)).collect(Collectors.toList());
     }
 
-    public List<PostFilterDTO> showNewsFeed(int uid, int pageNumber, int rowsNumber) {
+    List<PostFilterDTO> showNewsFeed(int uid, int pageNumber, int rowsNumber) {
         return dao.showNewsFeed(uid, pageNumber, rowsNumber);
     }
 
-    public List<PostFilterDTO> getPostsOfUser(int uid, int pageNumber, int rowsNumber) {
+    List<PostFilterDTO> getPostsOfUser(int uid, int pageNumber, int rowsNumber) {
         return dao.getPostsOfUser(uid, pageNumber, rowsNumber);
     }
 
-    public LikesDislikesDTO reactTo(int uid, int pid, String reaction) {
+    // removes old reaction after following visit
+    LikesDislikesDTO reactTo(int uid, int pid, String reaction) {
         User user = getVerifiedUserById(uid);
         Post post = getPostById(pid);
         PostReaction postReaction = new PostReaction();
@@ -190,7 +195,7 @@ public class PostService extends MasterService {
         }
     }
 
-    public List<UserIdNamesPhotoDTO> getUsersWhoReacted(int pid, String reaction) {
+    List<UserIdNamesPhotoDTO> getUsersWhoReacted(int pid, String reaction) {
         Post post = postRepository.findById(pid)
             .orElseThrow(() -> new NotFoundException("Post not found."));
         switch (reaction) {
